moved websocket code into separate files (outside of main.c)
