+++ main/src/websocket_handler.h
@@ -1,0 +1,8 @@
+#ifndef OPENPYROJET_WEBSOCKET_HANDLER_H
+#define OPENPYROJET_WEBSOCKET_HANDLER_H
+
+#include <esp_http_server.h>
+
+extern const httpd_uri_t websocketHandlerUri;
+
+#endif //OPENPYROJET_WEBSOCKET_HANDLER_H
