+++ main/src/wifi_connect.h
@@ -1,10 +1,6 @@
-/* Common functions for protocol examples, to establish Wi-Fi or Ethernet connection.
-
-   This example code is in the Public Domain (or CC0 licensed, at your option.)
-
-   Unless required by applicable law or agreed to in writing, this
-   software is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
-   CONDITIONS OF ANY KIND, either express or implied.
+/**
+ * Common functions for protocol examples, to establish Wi-Fi or Ethernet connection.
+ * This code is based on example code from the ESP IDF SDK.
  */
 
 #pragma once
